Forklift_Ticket_System/Backend/requirements.txt
django==5.1.3
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[token_blacklist]
python-decouple
psycopg2-binary
django-cors-headers
pillow
googletrans==4.0.0-rc1
django-redis
python-dotenv
django_extensions
pytz
